changed int to bigint in database.sql

+++ dataBase.sql
@@ -25,13 +25,13 @@ DROP TABLE IF EXISTS `booking`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `booking` (
-  `id` int(11) NOT NULL,
-  `id_room` int(11) NOT NULL,
-  `id_service` int(11) DEFAULT NULL,
+  `id` bigint(20) NOT NULL,
+  `id_room` bigint(20) NOT NULL,
+  `id_service` bigint(20) DEFAULT NULL,
   `arrival_date` datetime NOT NULL,
   `leaving_date` datetime NOT NULL,
-  `id_client` int(11) NOT NULL,
-  `human_amount` int(11) NOT NULL,
+  `id_client` bigint(20) NOT NULL,
+  `human_amount` bigint(20) NOT NULL,
   PRIMARY KEY (`id`),
   KEY `idroom_idx` (`id_room`),
   KEY `idservice_idx` (`id_service`),
@@ -59,15 +59,15 @@ DROP TABLE IF EXISTS `client`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `client` (
-  `id` int(11) NOT NULL,
-  `id_user` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
+  `id_user` bigint(20) NOT NULL,
   `name` varchar(45) NOT NULL,
   `surname` varchar(45) NOT NULL,
   `phone` varchar(45) NOT NULL,
   `email` varchar(45) NOT NULL,
   `passport_num` varchar(45) NOT NULL,
-  `count_visiting` int(11) NOT NULL DEFAULT '0',
-  `id_discount` int(11) DEFAULT NULL,
+  `count_visiting` bigint(20) NOT NULL DEFAULT '0',
+  `id_discount` bigint(20) DEFAULT NULL,
   PRIMARY KEY (`id`),
   KEY `iduser_idx` (`id_user`),
   KEY `iddiscount_idx` (`id_discount`),
@@ -93,9 +93,9 @@ DROP TABLE IF EXISTS `discount`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `discount` (
-  `id` int(11) NOT NULL,
-  `percent` int(11) NOT NULL,
-  `min_visiting_count` int(11) NOT NULL DEFAULT '1',
+  `id` bigint(20) NOT NULL,
+  `percent` bigint(20) NOT NULL,
+  `min_visiting_count` bigint(20) NOT NULL DEFAULT '1',
   `description` varchar(45) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -118,10 +118,10 @@ DROP TABLE IF EXISTS `feedback`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `feedback` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `mark` varchar(45) NOT NULL,
   `text` varchar(255) DEFAULT NULL,
-  `id_client` int(11) NOT NULL,
+  `id_client` bigint(20) NOT NULL,
   PRIMARY KEY (`id`),
   KEY `idclients_idx` (`id_client`),
   CONSTRAINT `idclients` FOREIGN KEY (`id_client`) REFERENCES `client` (`id`)
@@ -145,11 +145,11 @@ DROP TABLE IF EXISTS `hotel_room`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `hotel_room` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `status` varchar(10) NOT NULL,
-  `room_amount` int(11) NOT NULL DEFAULT '1',
-  `id_type` int(11) NOT NULL,
-  `id_client` int(11) DEFAULT NULL,
+  `room_amount` bigint(20) NOT NULL DEFAULT '1',
+  `id_type` bigint(20) NOT NULL,
+  `id_client` bigint(20) DEFAULT NULL,
   `price` double NOT NULL,
   PRIMARY KEY (`id`),
   KEY `idclient_idx` (`id_client`),
@@ -176,9 +176,9 @@ DROP TABLE IF EXISTS `role`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `role` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `role_name` varchar(45) NOT NULL,
-  `permission` int(11) DEFAULT NULL,
+  `permission` bigint(20) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
 /*!40101 SET character_set_client = @saved_cs_client */;
@@ -200,9 +200,9 @@ DROP TABLE IF EXISTS `room_type`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `room_type` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `name` varchar(45) NOT NULL,
-  `human_amount` int(11) NOT NULL,
+  `human_amount` bigint(20) NOT NULL,
   `description` varchar(255) DEFAULT NULL,
   PRIMARY KEY (`id`)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -225,7 +225,7 @@ DROP TABLE IF EXISTS `service`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `service` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `name` varchar(45) NOT NULL,
   `price` double NOT NULL DEFAULT '0',
   `description` varchar(255) DEFAULT NULL,
@@ -251,10 +251,10 @@ DROP TABLE IF EXISTS `user`;
 /*!40101 SET @saved_cs_client     = @@character_set_client */;
  SET character_set_client = utf8mb4 ;
 CREATE TABLE `user` (
-  `id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `login` varchar(45) NOT NULL,
   `password` varchar(45) NOT NULL,
-  `id_role` int(11) NOT NULL,
+  `id_role` bigint(20) NOT NULL,
   PRIMARY KEY (`id`),
   KEY `idrole_idx` (`id_role`),
   CONSTRAINT `idrole` FOREIGN KEY (`id_role`) REFERENCES `role` (`id`)
@@ -279,4 +279,4 @@ UNLOCK TABLES;
 /*!40101 SET COLLATION_CONNECTION=@OLD_COLLATION_CONNECTION */;
 /*!40111 SET SQL_NOTES=@OLD_SQL_NOTES */;
 
--- Dump completed on 2019-03-10 20:17:36
+-- Dump completed on 2019-03-10 20:17:36hotel_roomhotel_room
